+++ lib/zipkin-tracer/rack/zipkin-tracer.rb
@@ -13,11 +13,13 @@
 # limitations under the License.
 require 'finagle-thrift'
 require 'finagle-thrift/trace'
-require 'sucker_punch'
 require 'zipkin-tracer/config'
+require 'zipkin-tracer/tracer_factory'
 
 module ZipkinTracer
 
+  # This middleware reads Zipkin headers from the request and sets/creates a Trace.id usable by the rest of the app
+  # It will also send the trace to the Zipkin service using one of the methods configured.
   class RackHandler
     B3_REQUIRED_HEADERS = %w[HTTP_X_B3_TRACEID HTTP_X_B3_PARENTSPANID HTTP_X_B3_SPANID HTTP_X_B3_SAMPLED]
     B3_OPT_HEADERS = %w[HTTP_X_B3_FLAGS]
@@ -25,59 +27,32 @@ module ZipkinTracer
     def initialize(app, config = nil)
       @app = app
       @lock = Mutex.new
-
-      config = Config.new(app, config)
-      SuckerPunch.logger = config.logger
-      adapter = config.adapter
-
-      ::Trace.tracer = case adapter
-        when :json
-          require 'zipkin-tracer/zipkin_json_tracer'
-          ::Trace::ZipkinJsonTracer.new(json_api_host: config.json_api_host, traces_buffer: config.traces_buffer)
-        when :scribe
-          require 'zipkin-tracer/careless_scribe'
-          ::Trace::ZipkinTracer.new(CarelessScribe.new(config.scribe_server), config.scribe_max_buffer)
-        when :kafka
-          require 'zipkin-tracer/zipkin_kafka_tracer'
-          ::Trace::ZipkinKafkaTracer.new(zookeepers: config.zookeeper)
-        else
-          ::Trace::NullTracer.new
-      end
-
-      ip_format = config.adapter == :json ? :string : :i32
-      ::Trace.default_endpoint = ::Trace::Endpoint.make_endpoint(
-        nil, # auto detect hostname
-        config.service_port,
-        service_name(config.service_name),
-        ip_format
-      )
-      ::Trace.sample_rate = config.sample_rate
-
-      @config = config
+      @config = Config.new(app, config).freeze
+      TracerFactory.new.tracer(@config)
     end
 
     def call(env)
-      # skip certain requests
-      return @app.call(env) if filtered?(env) || !routable_request?(env)
-
-      whitelisted = force_sample?(env)
-      id = get_or_create_trace_id(env, whitelisted) # note that this depends on the sample rate being set
-      tracing_filter(id, env, whitelisted) { @app.call(env) }
+      zipkin_env = ZipkinEnv.new(env, @config)
+      with_trace_id(zipkin_env.trace_id) do
+        if !Trace.id.sampled? || !routable_request?(env)
+          @app.call(env)
+        else
+          trace!(zipkin_env) { @app.call(env) }
+        end
+      end
     end
 
     private
 
-    def record(annotation)
-      ::Trace.record(annotation)
-    # Nothing wonky that the tracer does should stop us from using the app!!!
-    # Usually is better to rescue StandardError but the socket layer can launch Errno kind of exceptions
-    rescue Exception
-      # Ignore low level errors
+    def with_trace_id(trace_id, &block)
+      Trace.push(trace_id)
+      yield
+    ensure
+      Trace.pop
     end
 
-    # Use the Domain environment variable to extract the service name, otherwise use the default config name
-    def service_name(default_name)
-      ENV["DOMAIN"].to_s.empty? ? default_name : ENV["DOMAIN"].split('.').first
+    def record(annotation)
+      Trace.record(annotation)
     end
 
     # If the request is not valid for this service, we do not what to trace it.
@@ -89,43 +64,28 @@ module ZipkinTracer
       false
     end
 
-    def annotate(env, status, response_headers, response_body)
+    def annotate_plugin(env, status, response_headers, response_body)
       @config.annotate_plugin.call(env, status, response_headers, response_body) if @config.annotate_plugin
     end
 
-    def filtered?(env)
-      @config.filter_plugin && !@config.filter_plugin.call(env)
-    end
-
-    def force_sample?(env)
-      @config.whitelist_plugin && @config.whitelist_plugin.call(env)
-    end
-
-    def tracing_filter(trace_id, env, whitelisted)
+    def trace!(zipkin_env, &block)
       synchronize do
-        ::Trace.push(trace_id)
         #if called by a service, the caller already added the information
-        add_request_information(env) unless called_with_zipkin_headers?(env)
-        record(::Trace::Annotation.new(::Trace::Annotation::SERVER_RECV, ::Trace.default_endpoint))
-        record(::Trace::Annotation.new('whitelisted', ::Trace.default_endpoint)) if whitelisted
+        trace_request_information(zipkin_env.env) unless zipkin_env.called_with_zipkin_headers?
+        record(Trace::Annotation.new(Trace::Annotation::SERVER_RECV, Trace.default_endpoint))
+        record(Trace::Annotation.new('whitelisted', Trace.default_endpoint)) if zipkin_env.force_sample?
       end
       status, headers, body = yield
     ensure
       synchronize do
-        annotate(env, status, headers, body)
-        record(::Trace::Annotation.new(::Trace::Annotation::SERVER_SEND, ::Trace.default_endpoint))
-        ::Trace.pop
+        annotate_plugin(zipkin_env.env, status, headers, body)
+        record(Trace::Annotation.new(Trace::Annotation::SERVER_SEND, Trace.default_endpoint))
       end
-      [status, headers, body]
     end
 
-    def add_request_information(env)
-      ::Trace.set_rpc_name(env['REQUEST_METHOD'].to_s.downcase) # get/post and all that jazz
-      record(::Trace::BinaryAnnotation.new('http.uri', env['PATH_INFO'], 'STRING', ::Trace.default_endpoint))
-    end
-
-    def called_with_zipkin_headers?(env)
-      B3_REQUIRED_HEADERS.all? { |key| env.has_key?(key) }
+    def trace_request_information(env)
+      Trace.set_rpc_name(env['REQUEST_METHOD'].to_s.downcase) # get/post and all that jazz
+      record(Trace::BinaryAnnotation.new('http.uri', env['PATH_INFO'], 'STRING', Trace.default_endpoint))
     end
 
     def synchronize(&block)
@@ -136,19 +96,55 @@ module ZipkinTracer
       @config.logger.error("Exception #{e.message} while sending Zipkin traces. #{e.backtrace}")
     end
 
-    def get_or_create_trace_id(env, whitelisted, default_flags = ::Trace::Flags::EMPTY)
-      trace_parameters = if called_with_zipkin_headers?(env)
-                           env.values_at(*B3_REQUIRED_HEADERS)
-                         else
-                           new_id = Trace.generate_id
-                           [new_id, nil, new_id, ('true' if whitelisted || Trace.should_sample?)]
-                         end
-      trace_parameters[3] = (trace_parameters[3] == 'true')
+    # Environment with Zipkin information in it
+    class ZipkinEnv
+      attr_reader :env
 
-      trace_parameters += env.values_at(*B3_OPT_HEADERS) # always check flags
-      trace_parameters[4] = (trace_parameters[4] || default_flags).to_i
+      def initialize(env, config)
+        @env    = env
+        @config = config
+      end
 
-      Trace::TraceId.new(*trace_parameters)
+      def trace_id(default_flags = Trace::Flags::EMPTY)
+        trace_parameters = if called_with_zipkin_headers?
+                             @env.values_at(*B3_REQUIRED_HEADERS)
+                           else
+                             new_id = Trace.generate_id
+                             [new_id, nil, new_id]
+                           end
+        trace_parameters[3] = should_trace?(trace_parameters[3])
+        trace_parameters += @env.values_at(*B3_OPT_HEADERS) # always check flags
+        trace_parameters[4] = (trace_parameters[4] || default_flags).to_i
+
+        Trace::TraceId.new(*trace_parameters)
+      end
+
+      def called_with_zipkin_headers?
+        B3_REQUIRED_HEADERS.all? { |key| @env.has_key?(key) }
+      end
+
+      def force_sample?
+        @config.whitelist_plugin && @config.whitelist_plugin.call(@env)
+      end
+
+      private
+
+      def current_trace_sampled?
+        rand < @config.sample_rate
+      end
+
+      def should_trace?(parent_trace_sampled)
+        if parent_trace_sampled  #A service upstream decided this goes in all the way
+          parent_trace_sampled == 'true'
+        else
+          force_sample? || current_trace_sampled? && !filtered?
+        end
+      end
+
+      def filtered?
+        @config.filter_plugin && !@config.filter_plugin.call(@env)
+      end
     end
+
   end
 end