+++ lib/zipkin-tracer/zipkin_tracer_base.rb
@@ -13,7 +13,6 @@ module Trace
 
     def initialize(options={})
       @traces_buffer = options[:traces_buffer] || raise(ArgumentError, 'A proper buffer must be setup for the Zipkin tracer')
-      @options = options
       reset
     end
 
@@ -27,9 +26,10 @@ module Trace
       when Annotation
         span.annotations << annotation
       end
+      count = current_count
+      set_current_count(count + 1)
 
-      @count += 1
-      if @count >= @traces_buffer || (annotation.is_a?(Annotation) && annotation.value == Annotation::SERVER_SEND)
+      if current_count >= @traces_buffer || (annotation.is_a?(Annotation) && annotation.value == Annotation::SERVER_SEND)
         flush!
         reset
       end
@@ -47,14 +47,26 @@ module Trace
 
     private
 
+    def spans
+      Thread.current[:zipkin_spans] ||= {}
+    end
+
+    def current_count
+      Thread.current[:zipkin_spans_count] ||= 0
+    end
+
+    def set_current_count(count)
+      Thread.current[:zipkin_spans_count] = count
+    end
+
     def get_span_for_id(id)
       key = id.span_id.to_s
-      @spans[key] ||= Span.new("", id)
+      spans[key] ||= Span.new("", id)
     end
 
     def reset
-      @count = 0
-      @spans = {}
+      Thread.current[:zipkin_spans] = {}
+      set_current_count(0)
     end
   end
 end